+++ ChangeLog.md
@@ -8,6 +8,7 @@
 - Change `StatefulServer` class definition to not add `Processes` to the effects by default
 - Add forgotten re-exports to `Concurrent`
 - Fix the `NFData` instance for `Pdu (Observer o)`
+- Remove the `Reader` effect from the action parameter of `withObservationQueue`
 
 ## 0.25.0 
 - Improve effect type aliases and module structure, [read the details here](./ChangeLog-Details-0.25.0.md).